backend/pyproject.toml
[project]
name = "alili-backend"
version = "1.0.0"
description = "Yoga pose recognition backend using FastAPI and MediaPipe"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.123.10",
    "google-genai>=1.0.0",
    "mediapipe>=0.10.21",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.0.0",
    "python-socketio>=5.15.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (FastAPI depends on this)
    "B904",   # raise without from (too noisy)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
